+++ src/components/NavBarNew/components/Account/AccountLoggedIn.tsx
@@ -1,4 +1,4 @@
-import { styles } from '../styles'
+import { styles } from '../../styles'
 
 // Hooks
 import { useWeb3React } from '@web3-react/core'
